LC contest + more Code refactoring
